Perform association tests on binary or quantitative traits

Main method for performing association tests on binary or quantitative traits. Performs either aggregate tests or single variant tests, with a wide range of available statistical methods (see 'Details' for more information). Various parameters allow for implementation of for example recessive/dominant analyses, MAF-weighted burden tests and permutation-based association tests. Variant and sample filters (e.g. maximum MAF, minimum number of carriers) can be specified. assocTest can be run either directly on a genoMatrix object or alternatively on a gdb object, in which case assocTest can loop through multiple variant sets by providing a varSetFile or varSetList as input. Moreover, assocTest can be run on sets of genes ('gene set burden'), for which gene burden scores have been generated using aggregate and stored in an aggregateFile. For running assocTest on a genoMatrix object see: assocTest-genoMatrix for details
for running assocTest on a gdb object see: assocTest-gdb for details for running assocTest on a aggregateFile object see: assocTest-aggregateFile for details

Value

The object returned is of class rvbResult or singlevarResult. The columns are described below: (note that some columns are specific to either rvbResult or singlevarResult

respectively)

unit: (rvbResult) name of the unit tested (from varSet or from genoMatrix metadata)

VAR_id: (singlevarResult) VAR_id of the respective variant.

cohort: Name of the cohort

varSetName: varSetName included in the provided varSet, if provided, or from the genoMatrix metadata.

name: if specified, contains the name argument specified in assocTest.

pheno: Phenotype tested, as specified by the pheno argument.

covar: Covariates included, as specified by the covar argument.

geneticModel: genetic model used, as specified by the geneticModel argument.

MAFweight: MAF-weighting applied, as specified by the MAFweights argument.

test: statistical test used, as specified by the test argument.

nvar: (rvbResult) number of variants included in the aggregate test.

caseMAC: (singlevarResult) Minor allele count among cases

ctrlMAC: (singlevarResult) Minor allele count among controls.

caseCarriers: (rvbResult) The number of cases that carry at least one variant among variants in the variant set.

ctrlCarriers: (rvbResult) The number of controls that carry at least one variant among variants in the variant set.

meanCaseScore: (rvbResult) Mean burden score among cases.

meanCtrlScore: (rvbResult) Mean burden score among controls.

caseN: Number of cases.

ctrlN: Number of controls.

caseCallRate: Variant call-rate in cases. For aggregate tests this is the average call-rate across variants.

ctrlCallRate: Variant call-rate in controls For aggregate tests this is the average call-rate across variants.

effectAllele: (singlevarResult) Allele to which the effect estimate refers.

otherAllele: (singlevarResult) Non-effect allele

effect: Effect estimate for specified statistical test. Note that SKAT tests, ACAT-v tests, SPA tests and negative binomial tests don't yield effect estimates.

effectSE: Standard error of effect estimate of specified statistical test. Note that SKAT tests, ACAT-v tests, SPA tests and negative binomial tests don't yield effect estimates.

effectCIlower: Lower confidence interval of effect estimate of specified statistical test. Note that SKAT tests, ACAT-v tests, SPA tests and negative binomial tests don't yield effect estimates.

effectCIupper: Upper confidence interval of effect estimate of specified statistical test. Note that SKAT tests, ACAT-v tests, SPA tests and negative binomial tests don't yield effect estimates.

OR: Odds-ratio for glm and firth tests.

P: P-value for specified statistical test.

Details

Aggregate tests For aggregate tests, the following tests are implemented:

binary/quantitative traits:

skat: SKAT test as implemented in the SKAT R package (Wu et. al, 2011).

skat_burden: Burden test as implemented in the SKAT R package (Wu et. al, 2011).

skato: SKAT-O (Optimal Unified Test) as implemented in the SKAT R package (Wu et. al, 2011).

acatv: ACAT-V test as implemented in the ACAT R package (Liu et al., 2019)

quantitative traits:

lm: linear model

binary traits:

firth: firth logistic regression (Firth, 1993). The maximum number of iterations can be specified using the maxitFirth parameter

glm: logistic regression

nbinom: Negative binomial test

skat_robust: robust SKAT test (robust in the presence of an unbalanced case/control ratio) as implemented in the SKAT R package (Zhao et al., 2020).

skat_burden_robust: robust burden test (robust in the presence of an unbalanced case/control ratio) as implemented in the SKAT R package (Zhao et al., 2020).

skato_robust: robust SKAT-O test (robust in the presence of an unbalanced case/control ratio) (Zhao et al., 2020)

acatvSPA: adjusted ACAT-v test, using a score test using saddlepoint approximation.

acatvfirth: adjusted ACAT-v test, using Firth's logistic regression to perform single variant tests.

Single variant tests

For single variants, the following tests are implemented:

quantitative traits:

lm: linear model

binary traits:

scoreSPA: score test with saddle point approximation.

firth: firth logistic regression (Firth, 1993). The maximum number of iterations can be specified using the maxitFirth parameter.

glm: logistic regression

References

Wu, M. C. et al. Rare-Variant Association Testing for Sequencing Data with the Sequence Kernel Association Test. The American Journal of Human Genetics 89, 82–93 (2011).

Liu, Y. et al. ACAT: A Fast and Powerful p Value Combination Method for Rare-Variant Analysis in Sequencing Studies. The American Journal of Human Genetics 104, 410–421 (2019).

Firth, D. Bias Reduction of Maximum Likelihood Estimates. Biometrika 80, 27–38 (1993).

Zhao, Z. et al. UK Biobank Whole-Exome Sequence Binary Phenome Analysis with Robust Region-Based Rare-Variant Test. The American Journal of Human Genetics 106, 3–12 (2020).