assoctest-aggregateFile

Run assocTest on a aggregateFile object. See the main assocTest page for details.

# S4 method for aggregateFile
assocTest(
  object,
  pheno,
  test,
  geneSet,
  gdb,
  cohort = "SM",
  name = "none",
  continuous = FALSE,
  covar = NULL,
  substractCovar = NULL,
  maxitFirth = 1000,
  keep = NULL,
  output = NULL
)

Arguments

object

a aggregateFile object (generated using the aggregate method).

pheno

field to test as response variable, the response variable can either be binary (0/1) or continuous. If the response variable is continuous set continuous to TRUE.

test

Vector of statistical tests to run, options include firth,glm,lm, and nbinom.

geneSet

a geneSetList or geneSetFile object

gdb

a gdb object, this will be used to load the cohort data.

cohort

Cohort name (present in the gdb)

name

Optional name of the analysis.

continuous

Is the response variable continuous? (TRUE/FALSE). Defaults to FALSE.

covar

Character vector of covariates, or list with multiple sets of covariates.

substractCovar

Covariate from which aggregate should be substracted. Useful when adjusting for total variant counts, by specifying the total variant count variable here, the aggregate score of the gene set tested will be substracted from the total count variable.

maxitFirth

Maximum number of iterations to use for estimating firth confidence intervals.

keep

vector of sample IDs to keep, defaults to NULL, in which case all samples are kept. Defaults to FALSE.

output

Output file path for results. Defaults to NULL, in which case results are not written but returned as a data.frame().