assocTest-gdb

Run assocTest on a gdb object. See the main assocTest page for details.

# S4 method for gdb
assocTest(
  object,
  pheno,
  test,
  cohort = "SM",
  varSet = NULL,
  VAR_id = NULL,
  name = "none",
  continuous = FALSE,
  singlevar = FALSE,
  covar = NULL,
  offset = NULL,
  geneticModel = "allelic",
  imputeMethod = NULL,
  MAFweights = "none",
  maxitFirth = 1000,
  checkPloidy = NULL,
  keep = NULL,
  output = NULL,
  methodResampling = NULL,
  resamplingFile = NULL,
  nResampling = 1000,
  outputResampling = FALSE,
  memlimitResampling = NULL,
  minCallrateVar = 0,
  maxCallrateVar = Inf,
  minCallrateSM = 0,
  maxCallrateSM = Inf,
  minMAF = 0,
  maxMAF = 1,
  minMAC = 0,
  maxMAC = Inf,
  minCarriers = 0,
  maxCarriers = Inf,
  minCarrierFreq = 0,
  maxCarrierFreq = Inf,
  memlimit = Inf,
  verbose = TRUE
)

Arguments

object

a gdb object

pheno

colData field to test as response variable, the response variable can either be binary (0/1) or continuous. If the response variable is continuous set continuous to TRUE. Multiple phenotypes can be specified, which will then be tested separately.

test

Vector of statistical tests to run, options include firth,glm,lm,nbinom,skat,skat_burden,skato,skat_fwe,skat_burden_fwe, skato_fwe,skat_robust,skato_robust,skat_burden_robust, acatv, acatvSPA. See assocTest for details.

cohort

If a valid cohort name is provided, then the uploaded data for this cohort is used to filter and annotate the genoMatrix object. If not specified, all samples in the gdb will be loaded.

varSet

a varSetFile or varSetList object, alternatively a vector of VAR_ids can be specified using the VAR_id parameter.

VAR_id

a vector of VAR_ids, alternatively the varSet parameter can be specified. If single variant tests are ran, the memlimit argument controls how many variants to analyze at a time.

name

Optional name for the analysis, defaults to "none".

continuous

If the response variable continuous? (TRUE/FALSE). Defaults to FALSE.

singlevar

Run single variant tests? (TRUE/FALSE). Defaults to FALSE, in which case aggregate tests are ran.

covar

Character vector of covariates, or a list of character vectors of covariates in which case each covariate set will be tested separately.

geneticModel

Which genetic model to apply? ('allelic', 'recessive' or 'dominant'). Defaults to allelic. Multiple geneticModels can be specified, in which case each will be analyzed separately.

imputeMethod

Which imputation method to apply? ('meanImpute' or 'missingToRef'). Defaults to meanImpute.

MAFweights

MAF weighting method. Currently Madsen-Browning ('mb') is implemented, by default no MAF weighting is applied. Multiple MAFweights can be specified (comma-delimited), in which case each will be analyzed separately.

maxitFirth

Maximum number of iterations to use for estimating firth confidence intervals. Defaults to 1000.

checkPloidy

Version of the human genome to use when assigning variant ploidy (diploid,XnonPAR,YnonPAR). Accepted inputs are GRCh37, hg19, GRCh38, hg38. If no value is provided then all variants are assigned the default ploidy of "diploid".

keep

Vector of sample IDs to keep, defaults to NULL, in which case all samples are kept.

output

Output file path for results. Defaults to NULL, in which case results are not written.

methodResampling

Which method to use for resampling? ('permutation' currently implemented) Defaults to NULL, in which case no resampling is performed.

resamplingFile

A resamplingFile object.

nResampling

Number of resamplings to perform if methodResampling is specified.

outputResampling

If TRUE or a filepath, results for each resampling are returned (or saved to the filepath). This can be useful if permutations are used to calculated to estimate correlations among genes for example. Defaults to FALSE in which case resampling is used to calculate resampled P-values, results for individual resamplings are not returned.

memlimitResampling

Maximum number of resamplings to perform at a time. Resampling generates a matrix of n x p, where n is the number of samples and p the number of resamplings thus, for large number of resamplings it can be more efficient to split the permutations in chunks of size memlimitResampling. Defaults to NULL in which case all permutations are performed.

minCallrateVar

Minimum genotype rate for variant retention.

maxCallrateVar

Maximum genotype rate for variant retention.

minCallrateSM

Minimum genotype rate for sample retention.

maxCallrateSM

Maximum genotype rate for sample retention.

minMAF

Minimum minor allele frequency for variant retention.

maxMAF

Maximum minor allele frequency for variant retention.

minMAC

Minimum minor allele count for variant retention.

maxMAC

Maximum minor allele count for variant retention.

minCarriers

Minimum carrier count for variant retention.

maxCarriers

Maximum carrier count for variant retention.

minCarrierFreq

Minimum carrier frequency for variant retention.

maxCarrierFreq

Maximum carrier frequency for variant retention.

memlimit

Maximum number of variants to load at once (if VAR_id is specified).

verbose

Should the function be verbose? (TRUE/FALSE), defaults to TRUE.

append

Relevant if the output parameter is not NULL. Should results be appended to output? Defaults to FALSE.