assocTest-genoMatrix

Run assocTest on a genoMatrix object. See the main assocTest page for details.

# S4 method for genoMatrix
assocTest(
  object,
  pheno,
  test,
  name = "none",
  continuous = FALSE,
  singlevar = FALSE,
  covar = NULL,
  offset = NULL,
  overwriteAggregate = TRUE,
  geneticModel = c("allelic", "recessive", "dominant"),
  imputeMethod = NULL,
  MAFweights = "none",
  maxitFirth = 1000,
  keep = NULL,
  output = NULL,
  append = FALSE,
  returnDF = FALSE,
  methodResampling = NULL,
  resamplingMatrix = NULL,
  nResampling = 1000,
  outputResampling = FALSE,
  memlimitResampling = NULL,
  minCallrateVar = 0,
  maxCallrateVar = Inf,
  minCallrateSM = 0,
  maxCallrateSM = Inf,
  minMAF = 0,
  maxMAF = 1,
  minMAC = 0,
  maxMAC = Inf,
  minCarriers = 0,
  maxCarriers = Inf,
  minCarrierFreq = 0,
  maxCarrierFreq = Inf,
  verbose = TRUE
)

Arguments

object

a genoMatrix object

pheno

colData field to test as response variable, the response variable can either be binary (0/1) or continuous. If the response variable is continuous set continuous to TRUE.

test

Vector of statistical tests to run, options include firth,glm,lm,scoreSPA,skat,skat_burden,skato,skat_fwe,skat_burden_fwe, skato_fwe,skat_robust,skato_robust,skat_burden_robust, acatv, acatvSPA. See assocTest for details.

name

Optional name for the analysis, defaults to "none".

continuous

Is the response variable continuous? (TRUE/FALSE). Defaults to FALSE.

singlevar

Run single variant tests? (TRUE/FALSE). Defaults to FALSE, in which case collapsing tests are ran.

covar

Character vector of covariates. These should be present in the colData slot of the genoMatrix.

overwriteAggregate

In case there is already an aggregate column in the colData of the genoMatrix (i.e. aggregate has been run on the genoMatrix), should it be overwitten? Defaults to TRUE.

geneticModel

Which genetic model to apply? ('allelic', 'recessive' or 'dominant'). Defaults to allelic.

imputeMethod

Which imputation method to apply? ('meanImpute' or 'missingToRef'). Defaults to meanImpute

MAFweights

Apply MAF weighting? Currently Madsen-Browning ('mb') is implemented. Defaults to 'none'.

maxitFirth

Maximum number of iterations to use for estimating firth confidence intervals.

output

Output file path for results. Defaults to NULL, in which case results are not written to disk, but returned as an rvatResult object.

append

Relevant if the output parameter is not NULL. Should results be appended to output? Defaults to FALSE.

methodResampling

Which method to use for resampling? ('permutation' currently implemented) Defaults to NULL, in which case no resampling is performed.

resamplingMatrix

Pre-calculated resampling matrix (n x p), where n = number of samples, and p number of resamplings. Can be generated using buildResamplingFile.

nResampling

Number of resamplings to perform if methodResampling is specified.

outputResampling

If TRUE or a filepath, results for each resampling are returned (or saved to the filepath). This can be useful if permutations are used to calculated to estimate correlations among genes for example. Defaults to FALSE in which case resampling is used to calculate resampled P-values, results for individual resamplings are not returned.

memlimitResampling

Maximum number of resamplings to perform at a time. Resampling generates a matrix of n x p, where n is the number of samples and p the number of resamplings thus, for large number of resamplings it can be more efficient to split the permutations in chunks of size memlimitResampling. Defaults to NULL in which case all permutations are performed.

minCallrateVar

Minimum genotype rate for variant retention.

maxCallrateVar

Maximum genotype rate for variant retention.

minCallrateSM

Minimum genotype rate for sample retention.

maxCallrateSM

Maximum genotype rate for sample retention.

minMAF

Minimum minor allele frequency for variant retention.

maxMAF

Maximum minor allele frequency for variant retention.

minMAC

Minimum minor allele count for variant retention.

maxMAC

Maximum minor allele count for variant retention.

minCarriers

Minimum carrier count for variant retention.

maxCarriers

Maximum carrier count for variant retention.

minCarrierFreq

Minimum carrier frequency for variant retention.

maxCarrierFreq

Maximum carrier frequency for variant retention.

verbose

Should the function be verbose? (TRUE/FALSE), defaults to TRUE.