The genoMatrix class for storing genotype data and associated sample and variant info.

The genoMatrix class is specifically designed to represent genotype data and associated sample and variant info. It extends the Bioconductor SummarizedExperiment class to accommodate genotype data.

Construct a genoMatrix

getGT(): Construct a genoMatrix, see getGT() for details.

Accessors

The accessors are fully described in SummarizedExperiment::SummarizedExperiment, and include assays(x), rowData(x), colData(x), metadata(x), dim(x), nrow(x), ncol(x), colnames(x) and rownames(x).

Getters

In the following code snippets, x is a genoMatrix object

getAF(x): Returns alternate allele frequencies. These will be equal to the minor allele frequences after applying flipToMinor().

getMAF(x): Returns variant minor allele frequencies.

getAC(x): Returns alternate allele counts. These will be equal to the minor allele counts after applying flipToMinor().

getMAC(x): Returns minor allele counts.

getCR(x, var=TRUE): Returns call-rates for variants or samples. var=TRUE by default, set to FALSE to return sample call-rates.

getNCarriers(x): Returns number of carriers of the alternate allele for each variant. Note that when geneticModel = 'recessive' it will return the number of homozygous carriers.

summariseGeno(x): Returns a per variant summary of genotype counts and hwe testing.

getCarriers(x, VAR_id=NULL,colDataFields=NULL,rowDataFields=NULL,groupBy=NULL): Return sample IDs for carriers of each of the variants in the genoMatrix. VAR_id can be specified to return for the specified subset of variants. colDataFields and rowDataFields can be specified include additional fields from colData(x) or rowData(x) in the output. The groupBy parameter can be set to calculate carrier frequency among groups (such as cohort or phenotype), multiple groups can be set. The aggregate parameter can be set to TRUE to return mean burden scores among the groupings.

note: Variant ploidy (diploid, XnonPAR, YnonPAR) are handled according to sample sex. Samples for which sex are not provided are excluded during AF calculation at non-diploid variants.

Subsetting

Subsetting a genoMatrix object is equivalent to subsetting a SummarizedExperiment object, and is fully described in fully described in SummarizedExperiment::SummarizedExperiment. Please see the example section for examples.

Recode

In the following code snippets, x is a genoMatrix object

flipToMinor(x): Function to flip genotype dosages such that all GT values represent minor allele counts.

recode(x, geneticModel,imputeMethod,weights,MAFweights): Returns a recoded genoMatrix object, genetic model, imputation, and weights can be recoded. See recode() for details.

updateGT(x, SM = NULL, anno = NULL): Safe replacement of the colData or rowData table within a genoMatrix with a new table.

Rare variant testing

In the following code snippets, x is a genoMatrix object

aggregate(): Returns per sample aggregate dosage for a genoMatrix object. The genoMatrix shouldn't contain missing values, use recode() to impute missing values. Aggregation is dependent on geneticModel, MAFweights and weights, all can be set using the recode() method. By default, an updated genoMatrix is returned with an aggregate field in colData. Set returnGT to FALSE to return a vector of aggregates.

assocTest(): Perform aggregate (burden) and single variant assocation test. See assocTest() for details.

See also