Load genotypes from a gdb.

Method to retrieve a genoMatrix for variants specified by varSet or a vector of VAR_ids, and samples as specified by the cohort parameter. The checkPloidy parameter can be set to GRCh37 (or hg19) or GRCh38 (or hg38) to assign variant ploidy (diploid,XnonPAR,YnonPAR).

getGT(
  object,
  varSet = NULL,
  VAR_id = NULL,
  ranges = NULL,
  cohort = NULL,
  checkPloidy = NULL,
  varSetName = "unnamed",
  unit = "unnamed",
  verbose = TRUE
)

# S4 method for gdb
getGT(
  object,
  varSet = NULL,
  VAR_id = NULL,
  ranges = NULL,
  cohort = NULL,
  checkPloidy = NULL,
  varSetName = "unnamed",
  unit = "unnamed",
  verbose = TRUE
)

Arguments

object

an object of class gdb

varSet

varSet object. If specified, the VAR_id, varSetName and unit parameters will be ignored.

VAR_id

Charaocter vector containing target VAR_id.

ranges

Extract variants within specified ranges. Ranges can be specified as a data.frame, including at least 'CHROM','start', and 'end' columns, or can be a GenomicRanges::GRanges object.

cohort

Optional use of cohort data previously uploaded to the gdb (see uploadCohort). If a valid cohort name is provided, then the uploaded data for this cohort is used to filter and annotate the returned genoMatrix object. If a dataframe is provided, then this is assumed to conform to the SM table constraints required for genoMatrix objects (see genoMatrix).

checkPloidy

Version of the human genome to use when assigning variant ploidy (diploid,XnonPAR,YnonPAR). Accepted inputs are GRCh37, hg19, GRCh38, hg38. If no value is provided then all variants are assigned the default ploidy of "diploid"

varSetName

Optional name for the set of variants, for example: 'missense or 'LOF' (ignored if varSet is specified.)

unit

Optional 'unit' name, for example: 'SOD1' or 'ENSG00000142168' (ignored if varSet is specified.)

verbose

Should the method be verbose? Defaults to TRUE.

Value

A genoMatrix object.