Class for handling and visualizing results generated using rvat.

The rvatResult class is specifically designed to represent association results generated using RVAT (with assocTest or geneSetAssoc for example). It extends the BioConductor DFrame class, and allows for basic operations such as subsetting and merging as well as visualization (manhattan, qqplot, forest plots) and downstream analyses (e.g. ACAT). Different type of results have their own subclasses (rvbResult, singlevarResult, gsaResult) that inherit from rvatResult.

Constructing

rvbResult(object, header = TRUE): Intitialize an rvbResult object, including singlevar results as generated by the assocTest method. object can be either a data.frame/DataFrame or a filepath pointing to the results.

singlevarResult(object, header = TRUE): Intitialize an singlevarResult object, including rvb results as generated by the assocTest method. object can be either a data.frame/DataFrame or a filepath pointing to the results.

gsaResult(object, header = TRUE): Intitialize an singlevarResult object, including rvb results as generated by the geneSetAssoc or assocTest methods. object can be either a data.frame/DataFrame or a filepath pointing to the results

readResults(path, header = TRUE, type = NULL, sep = "\t"): Alternatively, results can be read using readResults where type can be set to rvbResult/singlevarResult/gsaResult. If type is not set, it will be inferred from the input.

Subsetting

Subsetting an rvatResult object is equivalent to subsetting a DataFrame/DFrame object and is fully described in S4Vectors::DataFrame. DataFrame objects behave very similar to the base data.frame, with the most notable exceptions that row names are optional and it can hold alternative vectors such as run-length encoded vectors (S4Vectors::Rle). Please see the example section for examples.

Merging

merge(x, y, by): Merge an rvatResult object with a data.frame or DataFrame object. The by parameter specifies which variables to join by. For example, by = c("a" = "b") will match x$a to y$b. Join by multiple variable by providing a vector of length>1 to by.

Displaying

In the following code snippets, x is an rvatResult object

show(x): By the default the number of rows and columns will be displayed + the first five rows.

summary(x, asList = FALSE): Shows a summary of the contents of the rvatResult. Optionally the output can be stored as a list by setting asList = TRUE.

topResult(x, n=10): Show the top N most significant results (based on the P-value).

Visualization

qqplot: Plot a qqplot based on the P-values stored in an rvatResult object. See qqplot for details

manhattan: Generate a manhattan plot for an rvatResult object. See manhattan for details.

rvatViewer: Interactive visualization, see rvatViewer for details.

Downsteam analyses

geneSetAssoc: An rvbResult can be used as input for geneSetAssoc to perform gene set analyses. See geneSetAssoc for details.

ACAT: P-values in an rvbResult can be combined using the ACAT method. See ACAT for details.

Writing

writeResult(x, ...): Write an rvatResult to disk, see writeResult for details.

See also