Return variant summaries

Returns a per variant summary of genotype counts, frequencies, call-rates and hwe testing. Note, the gdb implementation is described here, summariseGeno can also be run directly on a genoMatrix object as described in the genoMatrix documentation.

summariseGeno(object, ...)

Arguments

object

a gdb object

cohort

If a valid cohort name is provided, then the uploaded data for this cohort is used to filter and annotate the genotypes If not specified, all samples in the gdb will be loaded.

varSet

a varSetList or varSetFile object.

VAR_id

A list of VAR_ids, alternatively the varSet parameter can be specified. If single variant tests are ran, the memlimit argument controls how many variants to analyze at a time.

pheno

colData field to test as response variable, although not used within this method, this can be useful to filter samples which have missing data for the response variable.

memlimit

Maximum number of variants to load at once (if VAR_id is specified).

geneticModel

Which genetic model to apply? ('allelic', 'recessive' or 'dominant'). Defaults to allelic.

checkPloidy

Version of the human genome to use when assigning variant ploidy (diploid,XnonPAR,YnonPAR). Accepted inputs are GRCh37, hg19, GRCh38, hg38. If no value is provided then all variants are assigned the default ploidy of "diploid".

keep

vector of sample IDs to keep, defaults to NULL, in which case all samples are kept.

output

Output file path for results. Defaults to NULL, in which case results are not written.

splitBy

Split variant summaries by labels indicated in the field specified by splitBy.

minCallrateVar

Minimum genotype rate for variant retention.

maxCallrateVar

Maximum genotype rate for variant retention.

minCallrateSM

Minimum genotype rate for sample retention.

maxCallrateSM

Maximum genotype rate for sample retention.

minMAF

Minimum minor allele frequency for variant retention.

maxMAF

Maximum minor allele frequency for variant retention.

minMAC

Minimum minor allele count for variant retention.

maxMAC

Maximum minor allele count for variant retention.

minCarriers

Minimum carrier count for variant retention.

maxCarriers

Maximum carrier count for variant retention.

minCarrierFreq

Minimum carrier frequency for variant retention.

maxCarrierFreq

Maximum carrier frequency for variant retention.

Value

Returns a data.frame with the following columns:

VAR_id: VAR_id of the respective variant.

AF: Allele frequency

callRate: callRate

geno0: Number of samples with genotype='0'. When geneticModel='allelic' or 'dominant' this is the number of individuals that are homozygous for the reference allele.

geno1: Number of samples with genotype='1'. When geneticModel='allelic' this is the number of individuals that are heterozygous for the reference allele. When geneticModel = 'dominant' this represents the number of individuals who carry at least one alternate allele. When geneticModel = 'recessive' this represents the number of individuals who are homozygous for the alternate allele.

geno2: When geneticModel = 'allelic', the number of individuals who are homozygous for the alternate allele.