Class to manage interactions with a varSetFile

An S4 class to manage interactions with varSetFiles. varSets can be generated from annotations using the buildVarSet() method. Specific units and/or annotations can be loaded using the getVarSet method. A varSetFile can be used as input in assocTest() to perform burden/single variant association tests on the varSets included in the varSetFile.

Build a varSetFile

buildVarSet(x, ...): Generate a varSetList or varSetFile() that stores weighted variant sets for use in association testing. This can be based on 1) annotations uploaded to the gdb or 2) a data.frame including annotations. See buildVarSet() for details.

Connect to a varSetFile

varSetFile(path, memlimit = 5000): Connect to a varSetFile object.

Getters

In the following code snippets, x is a varSetFile object.

getVarSet(x, unit, varSetName): Retrieve varSets for specified units and/or varSetNames. Output will be of class varSetList.

listUnits(x): Return a vector of all units included in the varSetFile

listVarSets(x): Return a vector of all varSetNames included in the varSetFile

Association testing

A varSetFile can be directly supplied to the assocTest() gdb method, using the varSet parameter. Association tests will then be performed for each varSet included in the varSetFile.

See also