Class to manage multiple varSets

An S4 class to manage varSets. varSets can be generated from annotations using the buildVarSet() method. Specific units and/or annotations can be retrieved using the getVarSet method. A varSetList can be used directly as input in assocTest() to perform burden/single variant association tests on the varSets included in the varSetList. The on-disk equivalent of a varSetList is varSetFile(), from which varSets can be retrieved in the same way as a varset list using the getVarSet method.

Build a varSetList

buildVarSet(x, ...): Generate a varSetList or varSetFile() that stores weighted variant sets for use in association testing. This can be based on 1) annotations uploaded to the gdb or 2) a data.frame including annotations. See buildVarSet() for details.

Getters

In the following code snippets, x is a varSetList object.

getVarSet(x, unit, varSetName): Retrieve varSets for specified units and/or varSetNames.

listUnits(x): Return a vector of all units included in the varSetList.

listVarSets(x): Return a vector of all varSetNames included in the varSetList.

Subsetting

A varSetList can be subsetted in the same way as a normal R list, however, getVarSet is the most convenient way to select varSets (see above). Some examples, where x is a varSetList object:

x[[i]]: Return the i'th varSet

x[i:j]: Return the i'th till j'th varSets

Association testing

A varSetList can be directly supplied to the assocTest() gdb method, using the varSet parameter. Association tests will then be performed for each varSet included in the varSetList.

Miscellaneous

In the following code snippets, x is a varSetList object:

collapseVarSetList(x, unit = "unnamed", varSetName = "unnamed): Merge all varSets into one varSet. All weights will be set to 1. Optionally, the unit name and varSetName can be specified. This method is mainly used when you want to load the genotypes for all variants in a varSetList using getGT().

write(x, file = "data", append = "FALSE"): Write the varSetList to disk, in the varSetFile format.

See also