Create a gdb file.

Creates a new gdb file and returns a connection object of type gdb-class. The gdb can be structured and populated using a provided vcf file. If no input variant file is provided then only an empty gdb is created.

buildGdb(
  output,
  vcf = c(),
  skipIndexes = FALSE,
  skipVarRanges = FALSE,
  overWrite = FALSE
)

Arguments

output

Path for output gdb file

vcf

Input vcf file used to structure and populate gdb. Warning this function makes the following of assumptions: 1) strict adherence to vcf format (GT subfield first element in genotype firelds), 2) multiallelic records have been split, 3) desired genotype QC has already been applied (DP,GQ filters), 4) GT values conform to the set 0/0,0/1,1/0,1/1,./.,0|0,0|1,1|0,1|1,.|.. Multiallelic parsing and genotype QC can be performed using vcftools and/or accompanying parser scripts included on the rvat github.

skipIndexes

Flag to skip generation of indexes for var and dosage table (VAR_id;CHROM, POS,REF,ALT). Typically only required if you plan to use gdbConcat to concatenate a series of separately generated gdb files before use

skipVarRanges

Flag to skip generation of ranged var table. Typically only required if you plan to use gdbConcat to concatenate a series of separately generated gdb files before use

overWrite

overwrite if output already exists? Defaults to FALSE, in which case an error is raised.