Create a gdb file.

Creates a new gdb file and returns a connection object of type gdb-class. The gdb can be structured and populated using a provided vcf file. If no input variant file is provided then only an empty gdb is created.

buildGdb(
  vcf,
  output,
  skipIndexes = FALSE,
  skipVarRanges = FALSE,
  overWrite = FALSE,
  genomeBuild = NULL,
  memlimit = 1000,
  verbose = TRUE
)

Arguments

vcf

Input vcf file used to structure and populate gdb. Warning this function makes the following of assumptions: 1) strict adherence to vcf format (GT subfield first element in genotype firelds), 2) multiallelic records have been split, 3) desired genotype QC has already been applied (DP,GQ filters), 4) GT values conform to the set 0/0,0/1,1/0,1/1,./.,0|0,0|1,1|0,1|1,.|.. Multiallelic parsing and genotype QC can be performed using vcftools and/or accompanying parser scripts included on the rvat github.

output

Path for output gdb file

skipIndexes

Flag to skip generation of indexes for var and dosage table (VAR_id;CHROM, POS,REF,ALT). Typically only required if you plan to use gdbConcat to concatenate a series of separately generated gdb files before use

skipVarRanges

Flag to skip generation of ranged var table. Typically only required if you plan to use gdbConcat to concatenate a series of separately generated gdb files before use

overWrite

overwrite if output already exists? Defaults to FALSE, in which case an error is raised.

genomeBuild

Optional genome build to include in the gdb metadata. If specified, it will be used to set ploidies (diploid, XnonPAR, YnonPAR) if the genome build is implemented in RVAT (currently: GRCh37, hg19, GRCh38, hg38).

memlimit

Maximum number of vcf records to parse at a time, defaults to 1000.

verbose

Should the function be verbose? (TRUE/FALSE), defaults to TRUE.

Examples

library(rvatData)
vcfpath <- rvat_example("rvatData.vcf.gz")
gdbpath <- tempfile()

# build a gdb from vcf. 
# the genomeBuild parameters stores the genome build in the gdb metadata
# this will be used to assign ploidies on sex chromosomes (diploid, XnonPAR, YnonPAR)
buildGdb(
  vcf = vcfpath,
  output = gdbpath,
  genomeBuild = "GRCh38"
)
#> 2024-11-20 14:40:05	Creating gdb tables
#> 25000 sample IDs detected
#> 2024-11-20 14:40:05	Parsing vcf records
#> 2024-11-20 14:40:32	Processing completed for 1000 records. Committing to db.
#> 2024-11-20 14:40:52	Processing completed for 1802 records. Committing to db.
#> 2024-11-20 14:40:52	Creating var table indexes
#> 2024-11-20 14:40:52	Creating SM table index
#> 2024-11-20 14:40:53	Creating dosage table index
#> 2024-11-20 14:40:53	Creating ranged var table
#> 2024-11-20 14:40:53	Complete
#> rvat gdb object
#>  Path: /var/folders/cl/wvc0rvjx4vd5rzt2_fhpmfth0000gp/T//Rtmp6OlT01/file77054b68b4c 

# for large vcfs, the memlimit parameter can be lowered
buildGdb(
  vcf = vcfpath,
  output = gdbpath,
  genomeBuild = "GRCh38",
  memlimit = 100,
  overWrite = TRUE
)
#> 2024-11-20 14:40:53	Creating gdb tables
#> 25000 sample IDs detected
#> 2024-11-20 14:40:53	Parsing vcf records
#> 2024-11-20 14:40:56	Processing completed for 100 records. Committing to db.
#> 2024-11-20 14:40:58	Processing completed for 200 records. Committing to db.
#> 2024-11-20 14:41:01	Processing completed for 300 records. Committing to db.
#> 2024-11-20 14:41:03	Processing completed for 400 records. Committing to db.
#> 2024-11-20 14:41:05	Processing completed for 500 records. Committing to db.
#> 2024-11-20 14:41:08	Processing completed for 600 records. Committing to db.
#> 2024-11-20 14:41:10	Processing completed for 700 records. Committing to db.
#> 2024-11-20 14:41:12	Processing completed for 800 records. Committing to db.
#> 2024-11-20 14:41:14	Processing completed for 900 records. Committing to db.
#> 2024-11-20 14:41:17	Processing completed for 1000 records. Committing to db.
#> 2024-11-20 14:41:19	Processing completed for 1100 records. Committing to db.
#> 2024-11-20 14:41:21	Processing completed for 1200 records. Committing to db.
#> 2024-11-20 14:41:23	Processing completed for 1300 records. Committing to db.
#> 2024-11-20 14:41:26	Processing completed for 1400 records. Committing to db.
#> 2024-11-20 14:41:28	Processing completed for 1500 records. Committing to db.
#> 2024-11-20 14:41:30	Processing completed for 1600 records. Committing to db.
#> 2024-11-20 14:41:32	Processing completed for 1700 records. Committing to db.
#> 2024-11-20 14:41:35	Processing completed for 1800 records. Committing to db.
#> 2024-11-20 14:41:35	Processing completed for 1802 records. Committing to db.
#> 2024-11-20 14:41:35	Creating var table indexes
#> 2024-11-20 14:41:35	Creating SM table index
#> 2024-11-20 14:41:35	Creating dosage table index
#> 2024-11-20 14:41:35	Creating ranged var table
#> 2024-11-20 14:41:35	Complete
#> rvat gdb object
#>  Path: /var/folders/cl/wvc0rvjx4vd5rzt2_fhpmfth0000gp/T//Rtmp6OlT01/file77054b68b4c 

# see ?gdb for more information on gdb-files, see ?concatGdb for concatenate gdb databases