Build a geneSetList or geneSetFile

Build a geneSetList or geneSetFile for use in gene set analyses (geneSetAssoc or assocTest-aggregateFile) Currently these can be build directly from GMT-files, data.frames and lists.

buildGeneSet(data = NULL, gmtpath = NULL, output = NULL, sep = "\t")

Arguments

data

Can be 1) data.frame where the first column includes the names of geneSets, the second column contains the genes included in the geneSet (comma-delimited). The third and fourth column are optional: the third column contains weights (comma-delimited), and the fourth column contains metadata. or 2) a list, where the names of the list represent the geneSet names and each element in the list contains a vector with the genes included in the respective geneset.

gmtpath

Path to a gmt-file

output

Optional output file path (output will be gz compressed text). Defaults to NULL, in which case a geneSetList is returned.

sep

Separator used in input file.