Connect and interact with a gdb

Compressed SQLite (".gdb") representation of genotype data and associated variant annotation and sample info tables These allow for rapid and memory-efficient of loading sample genotype data en associated metadata within R. The slots of the gdb class are inherited entirely from the RSQLite SQLiteConnection class. A host of RVAT methods described here allow for convenient querying and manipulation of a gdb, for complex queries users can also directly perform SQL queries on the gdb as exemplified in the examples and tutorials.

Slots

path

Path to gdb object

Connect

gdb(x): Connect to a gdb object (created using buildGdb()), where x is the filepath.

Build

In the following code snippets, x is a gdb object.

buildGdb(x): build a gdb directly from a vcf, see buildGdb() for details.

Getters

In the following code snippets, x is a gdb object.

getGT(x, ...): Retrieve genotype data from the gdb, returns a genoMatrix() object. See getGT() for details.

getAnno(x, ...): Get an annotation table from gdb. See the getAnno() documentation for details.

getCohort(x, ...): Get a cohort table from gdb. See the getCohort() documentation for details.

listCohort(x): List sample cohort tables that have been uploaded to the gdb.

listAnno(x): List variant info tables that have been uploaded to the gdb.

Upload and delete sample and variant info tables

In the following code snippets, x is a gdb object.

uploadCohort(x, ...): Upload cohort data tables to gdb. See the uploadCohort() documentation for details.

uploadAnno(x, ...): Upload variant annotation data into gdb. See the uploadAnno() documentation for details

mapVariants(x, ...): Map variants in the gdb onto features provided in a bed,gff/gtf or ranges file. See mapVariants() for details.

dropTable(x, name): Drop table with the name specified in name from gdb and clear from annotation / cohort metadata tables.

Subsetting & Merging

subSetGdb: Create a new gdb that is a subset of the input gdb. See subsetGdb() for details.

concatGdb: Function to concatenate gdb databases. See concatGdb() for details.

Exports

writeVcf: Convert a gdb to a vcf-file. See writeVcf() for details.

See also