geneSetAssoc

Perform self-contained or competitive gene set analyses.

geneSetAssoc(
  object,
  geneSet = NULL,
  scoreMatrix = NULL,
  cormatrix = NULL,
  condition = NULL,
  covar = NULL,
  test = c("lm", "glm", "mlm", "fisher", "ttest", "ztest", "ACAT"),
  threshold = NULL,
  Zcutoffs = NULL,
  INT = FALSE,
  scoreCutoffs = NULL,
  minSetSize = 1,
  maxSetSize = Inf,
  oneSided = TRUE,
  memlimit = 1000,
  REML = TRUE,
  ID = "unit",
  output = NULL
)

Arguments

object

an rvbResult object.

geneSet

a geneSetList or geneSetFile object.

scoreMatrix

a matrix (rows = genes, columns = features)

cormatrix

a correlation matrix with row and column names corresponding to the units in the rvbResult.

condition

Perform conditional analyses. Input can be a 1) a geneSetList or geneSetFile, in which case the genesets specified in the geneSet parameter will all be conditioned on the gene sets provided here. 2) a vector of gene set names present in geneSet, all genesets specified in the geneSetList/geneSetFile will be conditioned on the genesets specified here.

covar

a vector of covariates

test

Vector of tests to perform. Currently implemented tests are the competitive tests lm,mlm,fisher and self-contained tests ttest,ztest and ACAT.

threshold

A vector of thresholds for cutoff-based tests (fisher's exact test / glm).

Zcutoffs

A vector (length=2, minimum and maximum) of cutoffs to apply to the Z-scores. Z scores below/above these cutoffs will be set equal to the cutoff.

INT

Apply inverse normal transformation to Z-scores? Defaults to FALSE.

minSetSize

Exclude genesets with size < minSetSize

maxSetSize

Exclude genesets with size > maxSetSize

oneSided

Calculate a one-sided P-value? Defaults to TRUE.

memlimit

Maximum number of genesets to process in one go.

REML

Use REML for mixed linear models? Defaults to TRUE.

ID

ID column in the rvbResult that corresponds with the IDs used in the geneSetList. Defaults to 'unit'.

output

Optional: save results to specified path